backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
transformers
torch
scikit-learn
python-dotenv
sentence-transformers
passlib 
bcrypt 
pyjwt
python-multipart
